requirements.txt
python-dotenv
langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-core
faiss-cpu
youtube-transcript-api
